requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Calendar providers
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
msal==1.25.0

# LangGraph and LLM
langchain-core>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.30
langsmith>=0.0.60

# Semantic Search and Embeddings
openai>=1.0.0
chromadb>=0.4.0
numpy>=1.24.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
cryptography>=41.0.0